# packages/api-contract/openapi.yaml
openapi: 3.1.0
info:
  title: PersonalFin Application API
  version: 0.4.0
  description: >
    Application API contract. Authentication endpoints (sign-up, sign-in,
    sign-out, session) are served by Better Auth under /api/auth and are
    consumed through the Better Auth client; they are not part of this
    contract (ADR-0007).
security:
  - sessionCookie: []
paths:
...
  /financial-spaces/{spaceId}/tags:
    parameters:
      - $ref: '#/components/parameters/SpaceId'
    get:
      operationId: listTags
      summary: List the tags of the space
      responses:
        '200':
          description: Active tags first, then archived, by name.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagList'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/FinancialSpaceNotFound'
    post:
      operationId: createTag
      summary: Create a tag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
                - name
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 40
      responses:
        '201':
          description: The created tag.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/FinancialSpaceNotFound'
        '409':
          $ref: '#/components/responses/TagConflict'
  /financial-spaces/{spaceId}/tags/{tagId}:
    parameters:
      - $ref: '#/components/parameters/SpaceId'
      - name: tagId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      operationId: updateTag
      summary: Rename, archive, or unarchive a tag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
                - version
              properties:
                version:
                  type: integer
                  minimum: 1
                name:
                  type: string
                  minLength: 1
                  maxLength: 40
                archived:
                  type: boolean
      responses:
        '200':
          description: The tag after the change.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/TagNotFound'
        '409':
          $ref: '#/components/responses/TagConflict'
    delete:
      operationId: deleteTag
      summary: Permanently delete a never-used tag
      parameters:
        - name: version
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '204':
          description: Deleted.
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/TagNotFound'
        '409':
          $ref: '#/components/responses/TagConflict'
components:
  parameters:
...
    SpaceId:
      name: spaceId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: personalfin.session_token
      description: >
        Session cookie issued by Better Auth. Native clients send the same
        cookie through the Cookie header.
  responses:
    TagNotFound:
      description: The tag does not exist in this space (code TAG_NOT_FOUND).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    TagConflict:
      description: VERSION_CONFLICT, TAG_NAME_TAKEN, or TAG_IN_USE (archive instead).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ValidationFailed:
      description: The request is invalid (code VALIDATION_FAILED).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    FinancialSpaceNotFound:
      description: >
        The Financial Space does not exist or is not accessible to the user
        (code FINANCIAL_SPACE_NOT_FOUND). Both cases return 404 so that
        inaccessible spaces are not revealed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthenticated:
      description: No valid session was provided.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    Tag:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - archived
        - version
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        archived:
          type: boolean
        version:
          type: integer
          minimum: 1
    TagList:
      type: object
      additionalProperties: false
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          type: object
          additionalProperties: false
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Stable machine-readable error code.
            message:
              type: string
              description: Human-readable description; not localized.